anolis: sched/features: expand the space of sched features
ANBZ: #7717

With the increase of sched features, the available space decreases
gradually. In order to prevent features overflow, this patch expands
the space of sched features from 32 to 64.

Link: https://gitee.com/anolis/cloud-kernel/pulls/2502

+++ kernel/sched/sched.h
@@ -2278,7 +2278,7 @@ enum {
  * To support run-time toggling of sched features, all the translation units
  * (but core.c) reference the sysctl_sched_features defined in core.c.
  */
-extern const_debug unsigned int sysctl_sched_features;
+extern const_debug u64 sysctl_sched_features;
 
 #ifdef CONFIG_JUMP_LABEL
 #define SCHED_FEAT(name, enabled)					\
@@ -2308,7 +2308,7 @@ extern struct static_key sched_feat_keys[__SCHED_FEAT_NR];
  */
 #define SCHED_FEAT(name, enabled)	\
 	(1UL << __SCHED_FEAT_##name) * enabled |
-static const_debug __maybe_unused unsigned int sysctl_sched_features =
+static const_debug __maybe_unused u64 sysctl_sched_features =
 #include "features.h"
 	0;
 #undef SCHED_FEAT
